+++ serde_arrow/src/lib.rs
@@ -212,7 +212,8 @@ pub mod _impl {
                     pub use $arrow_array::types::{
                         ArrowPrimitiveType, Date32Type, Date64Type, Decimal128Type, Float16Type, Float32Type, Float64Type,
                         Int16Type, Int32Type, Int64Type, Int8Type, TimestampMicrosecondType, TimestampMillisecondType,
-                        TimestampNanosecondType, TimestampSecondType, Time64MicrosecondType, Time64NanosecondType, UInt16Type, UInt32Type,
+                        TimestampNanosecondType, TimestampSecondType, Time32MillisecondType, Time32SecondType, 
+                        Time64MicrosecondType, Time64NanosecondType, UInt16Type, UInt32Type,
                         UInt64Type, UInt8Type, ArrowDictionaryKeyType,
                     };
                     pub use $arrow_buffer::ArrowNativeType;